Fieldbook testing

+++ app/app/src/main/java/com/uu_uce/fieldbook/FieldbookAdapter.kt
@@ -85,10 +85,6 @@ class FieldbookAdapter(val activity: Activity, private val viewModel: FieldbookV
                         ViewGroup.LayoutParams.MATCH_PARENT
                     )
 
-                    customView.findViewById<CheckBox>(R.id.complete_box).isEnabled = false
-                    customView.findViewById<CheckBox>(R.id.complete_box).isVisible = false
-                    customView.findViewById<CheckBox>(R.id.complete_box).isClickable = false
-
                     // Add the title for the popup window
                     val windowTitle = customView.findViewById<TextView>(R.id.popup_window_title)
                     windowTitle.text = entry.location
@@ -124,7 +120,7 @@ class FieldbookAdapter(val activity: Activity, private val viewModel: FieldbookV
                         .setMessage("Are you sure you want to delete this entry?")
                         .setPositiveButton("YES") { _: DialogInterface, _: Int ->
                             viewModel.delete(entry)
-                            uri.toFile().delete()
+                            if (uri != Uri.EMPTY) uri.toFile().delete()
                         }
                         .setNegativeButton("NO") { _: DialogInterface, _: Int ->
